added support for CUTTLEFISH input files: .cfseg
